requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.5.1
attrs==20.3.0
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flake8==3.9.0
Flask==1.1.2
flask-restx==0.3.0
identify==2.1.3
importlib-metadata==3.7.3
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.9
pluggy==0.13.1
pre-commit==2.11.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.31.0
zipp==3.4.1
